+++ Application/ResearchDataManagementPlatform/WindowManagement/ContentWindowTracking/Persistence/PersistableObjectCollectionDockContent.cs
@@ -12,7 +12,6 @@ using Rdmp.Core.CommandExecution;
 using Rdmp.Core.Curation.Data.Dashboarding;
 using Rdmp.UI;
 using Rdmp.UI.ItemActivation;
-using Rdmp.UI.ItemActivation.Emphasis;
 using Rdmp.UI.Refreshing;
 using Rdmp.UI.TestsAndSetup.ServicePropogation;
 using ResearchDataManagementPlatform.WindowManagement.ExtenderFunctionality;