+++ Rdmp.UI/CohortUI/ExtractableCohortUI.cs
@@ -16,7 +16,6 @@ using Rdmp.Core.Providers;
 using Rdmp.UI.Collections;
 using Rdmp.UI.Copying;
 using Rdmp.UI.ItemActivation;
-using Rdmp.UI.ItemActivation.Emphasis;
 using Rdmp.UI.ScintillaHelper;
 using Rdmp.UI.SimpleControls;
 using Rdmp.UI.SimpleDialogs;