+++ Rdmp.UI/CommandExecution/AtomicCommands/ExecuteCommandUnpin.cs
@@ -8,7 +8,6 @@ using Rdmp.Core.CommandExecution;
 using Rdmp.Core.CommandExecution.AtomicCommands;
 using Rdmp.Core.Curation.Data;
 using Rdmp.UI.ItemActivation;
-using Rdmp.UI.ItemActivation.Emphasis;
 
 namespace Rdmp.UI.CommandExecution.AtomicCommands
 {