+++ Rdmp.UI/SubComponents/CohortCompilerUI.cs
@@ -27,7 +27,6 @@ using Rdmp.Core.QueryCaching.Aggregation;
 using Rdmp.UI.Collections;
 using Rdmp.UI.Icons.IconProvision;
 using Rdmp.UI.ItemActivation;
-using Rdmp.UI.ItemActivation.Emphasis;
 using Rdmp.UI.Refreshing;
 using Rdmp.UI.SimpleDialogs;
 using Rdmp.UI.SubComponents.EmptyLineElements;