+++ CHANGELOG.md
@@ -11,11 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Cloning an Extraction Configuration no longer expands clone and names the new copy "Clone of [..]" (previously name was a guid)
 - Select object dialog now display a maximum of 1000 objects (prioritising your search text)
+- Logging tasks are now case insensitive
 
 ### Fixed
 
 - Fixed Console input in CLI when running under Linux
-
+- Fixed issue where parallel checks could fail due to UI cross thread access
+- Fixed bugs in DLE when loading tables with dodgy column names (e.g. `[My Group by lolz]`)
+- 
 ...
 
 ### Added
@@ -28,7 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cohort Query Bulder will now connect to the database containing the data rather than the users default database when querying data on a single database
 - Flat file Attachers now process files in alphabetical order (case insensitive) when Pattern matches multiple files (previously order was arbitrary / OS defined)
 - Extraction source now specifies database to connect to when a dataset exists in a single database (previously connected to users default server e.g. master)
-- Updated to latest version of [FAnsiSql] (0.10.11) for Postgres support
+- Updated to latest version of [FAnsiSql] (0.10.12) for Postgres support
 - 
 ### Fixed